chore: reshuffle files, loadtest is now the main module
