+++ Cargo.toml
@@ -1,8 +1,25 @@
 [package]
-name = "scale-testing"
+name = "loadtest"
 version = "0.1.0"
 edition = "2021"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[workspace]
 
 [dependencies]
+anyhow = "1"
+chrono = "0.4"
+clap = { version = "4", features = ["derive", "env"] }
+goose = "=0.17.3-dev"
+goose-eggs = "0.5.3-dev"
+humantime = "2"
+log = "0.4"
+openid = "0.14"
+reqwest = "0.12"
+tokio = { version = "1.38.0", features = ["sync"] }
+
+[patch.crates-io]
+#goose = { path = "../../goose" }
+#goose-eggs = { path = "../../goose-eggs" }
+
+goose = { git = "https://github.com/ctron/goose", branch = "feature/baseline_1" }
+goose-eggs = { git = "https://github.com/ctron/goose-eggs", branch = "feature/uptick_deps_1" }
