--- loadtests/.gitignore
@@ -1,2 +1,0 @@
-report.*
-baseline.json
