+++ replicator/Cargo.toml
@@ -1,0 +1,6 @@
+[package]
+name = "scale-testing"
+version = "0.1.0"
+edition = "2021"
+
+[workspace]
